% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11example-package.R
\docType{package}
\name{cpp11example-package}
\alias{cpp11example}
\alias{cpp11example-package}
\title{cpp11example: Minimal example exploring memory management of external (C++) objects via R's cpp11 interface.}
\description{
Minimal example exploring memory management of external (C++) objects via R's cpp11 interface.
}
\author{
\strong{Maintainer}: Drew Herren \email{drewherrenopensource@gmail.com}

}
