% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{createExample}
\alias{createExample}
\title{Create an \code{Example} object}
\usage{
createExample(vec_size)
}
\arguments{
\item{vec_size}{Size of the C++ vector of \code{ObjectData} elements}
}
\value{
\code{Example} object
}
\description{
Create an \code{Example} object
}
