% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{Example}
\alias{Example}
\title{Example R object}
\description{
R object wrapping an external pointer to a C++ class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{example_ptr}}{External pointer to a C++ Example class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Example-new}{\code{Example$new()}}
\item \href{#method-Example-finalize}{\code{Example$finalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Example-new"></a>}}
\if{latex}{\out{\hypertarget{method-Example-new}{}}}
\subsection{Method \code{new()}}{
Create a new Example object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Example$new(vec_size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vec_size}}{Size of the C++ vector of \code{ObjectData} elements}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Example} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Example-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Example-finalize}{}}}
\subsection{Method \code{finalize()}}{
Code to run when Example object is removed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Example$finalize()}\if{html}{\out{</div>}}
}

}
}
